# apps/core-ledger-api/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Core Ledger API",
    "description": "Financial Ledger REST API for institutional financial clients",
    "contact": {
      "name": "Core Ledger Team",
      "email": "support@coreledger.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7109",
      "description": "Local development (HTTPS)"
    },
    {
      "url": "http://localhost:5071",
      "description": "Local development (HTTP)"
    }
  ],
  "paths": {
...
    "/api/historicos-indexadores": {
      "post": {
        "tags": [
          "HistoricosIndexadores"
        ],
        "operationId": "CreateHistoricoIndexador",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CreateHistoricoIndexadorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.HistoricoIndexadorDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/historicos-indexadores/{id}": {
      "delete": {
        "tags": [
          "HistoricosIndexadores"
        ],
        "operationId": "DeleteHistoricoIndexador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/indexadores": {
      "get": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "GetAllIndexadores",
        "parameters": [
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Tipo",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Periodicidade",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Fonte",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ativo",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "ImportacaoAutomatica",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.Models.PagedResult`1[[CoreLedger.Application.DTOs.IndexadorDto, CoreLedger.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "CreateIndexador",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.CreateIndexadorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.IndexadorDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/indexadores/{id}": {
      "get": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "GetIndexadorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.IndexadorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "UpdateIndexador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoreLedger.Application.DTOs.UpdateIndexadorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.DTOs.IndexadorDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "DeleteIndexador",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/indexadores/{id}/historico": {
      "get": {
        "tags": [
          "Indexadores"
        ],
        "operationId": "GetIndexadorHistorico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dataFim",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoreLedger.Application.Models.PagedResult`1[[CoreLedger.Application.DTOs.HistoricoIndexadorDto, CoreLedger.Application, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null]]"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoreLedger.Application.DTOs.CreateHistoricoIndexadorDto": {
        "required": [
          "dataReferencia",
          "indexadorId",
          "valor"
        ],
        "type": "object",
        "properties": {
          "indexadorId": {
            "type": "integer",
            "format": "int32"
          },
          "dataReferencia": {
            "type": "string",
            "format": "date-time"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "fatorDiario": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "variacaoPercentual": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "fonte": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "importacaoId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CoreLedger.Application.DTOs.CreateIndexadorDto": {
        "required": [
          "codigo",
          "nome",
          "periodicidade",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "codigo": {
            "maxLength": 20,
            "minLength": 1,
            "pattern": "^[A-Z0-9]+$",
            "type": "string"
          },
          "nome": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "tipo": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.IndexadorTipo"
          },
          "fonte": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "periodicidade": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Periodicidade"
          },
          "fatorAcumulado": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataBase": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "urlFonte": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "importacaoAutomatica": {
            "type": "boolean"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.HistoricoIndexadorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "indexadorId": {
            "type": "integer",
            "format": "int32"
          },
          "indexadorCodigo": {
            "type": "string",
            "nullable": true
          },
          "dataReferencia": {
            "type": "string",
            "format": "date-time"
          },
          "valor": {
            "type": "number",
            "format": "double"
          },
          "fatorDiario": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "variacaoPercentual": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "fonte": {
            "type": "string",
            "nullable": true
          },
          "importacaoId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.IndexadorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.IndexadorTipo"
          },
          "tipoDescricao": {
            "type": "string",
            "nullable": true
          },
          "fonte": {
            "type": "string",
            "nullable": true
          },
          "periodicidade": {
            "$ref": "#/components/schemas/CoreLedger.Domain.Enums.Periodicidade"
          },
          "periodicidadeDescricao": {
            "type": "string",
            "nullable": true
          },
          "fatorAcumulado": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataBase": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "urlFonte": {
            "type": "string",
            "nullable": true
          },
          "importacaoAutomatica": {
            "type": "boolean"
          },
          "ativo": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ultimoValor": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "ultimaData": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "historicoCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Application.DTOs.UpdateIndexadorDto": {
        "required": [
          "nome"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "fonte": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "fatorAcumulado": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataBase": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "urlFonte": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "importacaoAutomatica": {
            "type": "boolean"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CoreLedger.Domain.Enums.IndexadorTipo": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "CoreLedger.Domain.Enums.Periodicidade": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.\n\nExample: \"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}